Fix events spanning multiple months showing up multiple times in agenda.
Use Shared.isSameDate() in also places I was too lazy to also compare the year/month.
Fix performance with events spanning multiple months/years since it iterated every single day in the duration. We now iterate over the visible days in the month and check if that date is within the duration of the event. So events*eventDays*visibleDays O^3 => events*visibleDays O^2. In those equations, visibleDays is always 7*6=42 so performance is actually linear to the amount of events.