Create a new signal for when a date is selected in the monthview.
Previously, by depending on the onChanged signal, it did not fire if the selectedDate hadn't changed. So if you scrolled down in the agenda after selecting a date, then reselected the same date, it would not scroll back to the selection. This also might fix the scroll offset in the agenda upon loading the calendars. If so, we can retry showing events earlier from the current month if it always scrolls to the current date correctly.