Meteogram label fix. Deferred requests. Single column layout prep.
Fix preciptation labels (graph has the property, not meteogramView).
Create a variable to toggle the single column layout (that will expand to thes screen's height when both the agenda and the calendar is visible).
Change the precipation color to be more transparent.
Use the deferred pattern when calling for an update to the weather/events. We were previously calling it ~20 something times when the applet is first expanded which would each fire off requests. This would also cause the tooltip to jitter as it was constantly being updated as responses came in.
Log the calendar id when recieving events.
Enabled WeatherCanada (even though it's still not 100% done).
Use the AppletConfig pattern to parse plasmoid.configuration values (and to store hardcoded "config" values that aren't ready for serialization).