Use Label/NativeRendering for event summaries.
This should hopefully fix whatever weird bug is causing issue #60 in
Antergos/Arch with Qt 5.8. Text defaults to QtRendering, but emoji's
are being properly rendered in taskmanager which uses Label. Label uses
a C++ model, but logging the text with charCodeAt showed the same values
so this must be it.