Add a work in progress C++ plugin for volume feedback.
We need this plugin to stream the output of the python files since the executable dataengine only updates the stdout when the process is finished. It's not worth spamming pulseaudio and top with new audio stream monitors 10 times per second for a subpar effect.

Packagers should not build this this plugin, as I will be making changes to it without warning.