Add button to refresh the calendar list manually.
Previously we automatically updated the list when the config was opened,
but that caused bugs that were sort of hard to reproduce and even harder
to communicate how to fix the problem (by forcing a refresh by changing tabs...ew).
So might as well just add a button. This list doesn't really update often enough we
need to fetch it very often.