Copy the fixed width code from digitalclock.
We need to be able to support QML RichText formatting.

So first we use digitalclocks code that replaces hours/minutes/seconds with the widest number character.
I probably need to also replace milliseconds and ignore h/m/s characters inside formatting but that's too much work to do right now.
After that, we use digitalclock's code to use fontMetrics.advancedWidth on the replaced timeFormat.

Now instead of outputing the formatted text to sizehelper.text, we instead give it to another hidden Label.
We need to do this in order to calculate the width of the rendered RichText.

We do this for both timeFormat lines, then grab the max() if both lines are visible.

We now depend on QtQuick 2.4 in order to use FontMetrics.