Fix KHolidays events not appearing when viewing other months
CalendarBackend only fetches the events for the current month.
We need to force it to fetch events for a specfic month.

Right now, we're just telling it to set the CalendarBackend's month to
whatever month the day in the middle of the range we're requesting is.
A better solution would be to update the "displayedDate" in the date
iterator when we iterate to a different month.