Use Async.parallel to fetch gcal events.
If the calendarIdList contains an invalid id, then it will error and
the events will not be displayed. We need to detect deleted calendars
and deselet/remove them from calendarIdList. A warning notification
will be needed as well to inform the user. We will need to be sure
that the calendar was deleted and not that the user is offline or gcal
is down.