If all day event, populate end date with day before.
The end date is shifted forward in getEndObj() if it's still an
all day event when finished editing.

Also populate with time as 9-10am, which will be striped by
dateString() if All Day is not unchecked.

With these behaviours, if a user unchecks All Day, the end date is
useful, as it's no longer the next day at midnight.