Move logic for the dateMin/dateMax range to AgendaModel.
It should probably be in EventModel, but we need to get
AgendaModel.showNextNumDays, and we need access to
timeModel.currentTime to get today.

We simplify calculating the range by grabbing events +- 7 days from
the start/end of the month. So we will now always fetch events from
days we crop out.

We could move MonthView's daysModel calculation to get the exact
number of days shown before the start of the month, but it's not that
important.