Add EditTaskForm based EditEventForm (not finished)
Apparently GoogleTasks due date doesn't provide the time in the API.

We hide the functional CheckBox in AgendaTaskItem and show a disabled
copy in the EditTaskForm. The checkbox is aligned with the other form
icons, but isn't aligned with the other checkboxes.

We also need to set LinkRect.implicitHeight to contents.implicitHeight
to avoid a bug where the height stays huge after closing the EditForm.