Add konsolekalendar.py script to find the uid of an event
The PlasmaCalendar API only provides the summary, description, and
date so we need to find the closest event.

We need to set LC_TIME=C to have a consistent date format to parse.
I've submitted a patch to change the CSV output to use the ISO format.

https://invent.kde.org/pim/akonadi-calendar-tools/-/merge_requests/1

Eventually I will need to update to parse ISO formats, and fallback to
the LC_TIME=C formats to support older versions.