Change the client secret and refactor google api session config keys.
Change the config keys from snake_case to camelCase since we're
already making breaking changes by changing the client secret. We
still keep the snake_case access_token in order for migration purposes
so we can inform the user to relog when they update.

Store the client id+secret on login so we can periodically rotate it
without immediately breaking the user's session on update.

Also remove the device_code login config keys and code that was
removed in v62.