Play around with QQC2.ScrollView
A ListView inside a QQC2.ScrollView is better than QQC1, as when items
are added to the beginning of the model, the content below it does not
shift downwards. This means we can probably create an infinite scroll
above and below. The issue right now however is that ListView
scrolling is slow since delegate creation is slow.