Only store the filename in enabledCalendarPlugins (Issue #204)
Since qt5's /.../plugins/plasmacalendarplugins/ dir has a different
path in Ubuntu, Arch, and Fedora, we need to store just the filename.
If we don't we'd need to specify all filepaths (we already stored
Ubuntu and Arch) as the default config value.

Also document plugin directory paths in a few distros.