Fix Requests.js to fix weather not updating on connect (Issue #218)
Apparently XMLHttpRequest.onerror is called when there's no internet.
Also it does not have an "e" error argument.
Finally the onerror only gave the error message in the callback, it
didn't pass the data or xhr argument which means the logic code higher
up couldn't check if xhr.status == 0.

xhr.statusText is useless as it only shows "" or "OK" or "Not Found".