add quick tweaks to identify past events
With multiple calendars and dozens of meetings/events per calendar,
it is hard to distinguish past events from current day/future events.
Alongside the *inProgress state, a isPast property is added for events
that is used to place a tickmark ✓ for past events and a ■ for
current/future events in the calendar view tooltip.

The theme for the event list items with an end date in the past is
set to disabled, with the same tickmark prepended to the summary
so they can be ignored faster.

The event button, if any, is flattened for the same reason.