Updated russian translations by @algri14
Modified by @Zren, from @algri14's patch-6 branch:
https://github.com/algri14/plasma-applet-eventcalendar/commit/7fdff340af709bd2fe2ba981722c9c4d3363f689

@Zren: In the "Widget has been updated[...]" message I did not put add
the extra period between "[...]войдите. снова." which Google translate
shows as:
"войдите снова." => "and sign back in."
"войдите. снова." => "and sign in again."

Lastly there is a patch-7 branch with just the modified "General"
message, but is missing some of the empty translations from patch-6.
So I assumed that patch-6 is better than patch-7.
https://github.com/Zren/plasma-applet-eventcalendar/commit/a69f7299eb47b21e4362b30d6beefd3ee2eb9793