Fix restarted Plasma having incorrect CWD
Calling kstart5 from the CWD/repo of this project leads to all
newly opened programs will also run with CWD of the current project
path. In particular, opening konsole opens the shell directly in
the checkout of the repository.

Fix this by ensuring that we initiate the restart from $HOME.